move things around so we can implement AdditionOps in the right place and fix cyclical dependencies
