+++ db4-graph/Cargo.toml
@@ -15,7 +15,5 @@ edition.workspace = true
 boxcar.workspace = true
 storage.workspace = true
 raphtory-api.workspace = true
-raphtory-storage.workspace = true
 raphtory-core.workspace = true
-db4-common.workspace = true
 parking_lot.workspace = true
