+++ db4-storage/src/lib.rs
@@ -5,10 +5,10 @@ use std::{
 };
 
 use crate::{
+    error::DBV4Error,
     pages::GraphStore,
     segments::{edge::EdgeSegmentView, node::NodeSegmentView},
 };
-use db4_common::{LocalPOS, error::DBV4Error};
 use parking_lot::{RwLockReadGuard, RwLockWriteGuard};
 use raphtory_api::core::{
     entities::{
@@ -19,7 +19,7 @@ use raphtory_api::core::{
 };
 use segments::{edge::MemEdgeSegment, node::MemNodeSegment};
 
-pub mod loaders;
+// pub mod loaders;
 pub mod pages;
 pub mod persist;
 pub mod properties;
@@ -236,4 +236,79 @@ pub trait NodeRefOps<'a>: Copy + Clone + Send + Sync {
     fn c_prop(self, prop_id: usize) -> Option<Prop>;
 
     fn t_prop(self, prop_id: usize) -> Self::TProps;
+}
+
+pub mod error {
+    use std::{path::PathBuf, sync::Arc};
+
+    use raphtory_api::core::entities::properties::prop::PropError;
+    use raphtory_core::utils::time::ParseTimeError;
+
+    #[derive(thiserror::Error, Debug)]
+    pub enum DBV4Error {
+        #[error("External Storage Error {0}")]
+        External(#[from] Arc<dyn std::error::Error + Send + Sync>),
+        #[error("IO error: {0}")]
+        IO(#[from] std::io::Error),
+        #[error("Serde error: {0}")]
+        Serde(#[from] serde_json::Error),
+        // #[error("Load error: {0}")]
+        // LoadError(#[from] LoadError),
+        #[error("Arrow-rs error: {0}")]
+        ArrowRS(#[from] arrow_schema::ArrowError),
+        #[error("Parquet error: {0}")]
+        Parquet(#[from] parquet::errors::ParquetError),
+
+        #[error("Property error: {0}")]
+        PropError(#[from] PropError),
+        #[error("Empty Graph: {0}")]
+        EmptyGraphDir(PathBuf),
+        #[error("Failed to parse time string")]
+        ParseTime {
+            #[from]
+            source: ParseTimeError,
+        },
+        // #[error("Failed to mutate: {0}")]
+        // MutationError(#[from] MutationError),
+        #[error("Unnamed Failure: {0}")]
+        GenericFailure(String),
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
+#[repr(transparent)]
+pub struct LocalPOS(pub usize);
+
+impl LocalPOS {
+    pub fn as_vid(self, page_id: usize, max_page_len: usize) -> VID {
+        VID(page_id * max_page_len + self.0)
+    }
+
+    pub fn as_eid(self, page_id: usize, max_page_len: usize) -> EID {
+        EID(page_id * max_page_len + self.0)
+    }
+}
+
+impl From<usize> for LocalPOS {
+    fn from(pos: usize) -> Self {
+        Self(pos)
+    }
+}
+
+pub fn calculate_size_recursive(path: &Path) -> Result<usize, std::io::Error> {
+    let mut size = 0;
+    if path.is_dir() {
+        for entry in std::fs::read_dir(path)? {
+            let entry = entry?;
+            let path = entry.path();
+            if path.is_dir() {
+                size += calculate_size_recursive(&path)?;
+            } else {
+                size += path.metadata()?.len() as usize;
+            }
+        }
+    } else {
+        size += path.metadata()?.len() as usize;
+    }
+    Ok(size)
 }
