+++ db4-storage/src/persist/strategy.rs
@@ -1,4 +1,3 @@
-
 use std::ops::DerefMut;
 
 use crate::segments::{
