+++ db4-storage/src/segments/mod.rs
@@ -1,20 +1,18 @@
 use std::{collections::hash_map::Entry, fmt::Debug, sync::Arc};
 
 use bitvec::{order::Msb0, vec::BitVec};
-use db4_common::LocalPOS;
 use either::Either;
-use raphtory::{
-    core::{
-        entities::{
-            nodes::node_store::PropTimestamps,
-            properties::{tcell::TCell, tprop::TPropCell},
-        },
-        storage::timeindex::TimeIndexEntry,
+use raphtory_api::core::entities::properties::{meta::Meta, prop::Prop};
+use raphtory_core::{
+    entities::{
+        nodes::node_store::PropTimestamps,
+        properties::{tcell::TCell, tprop::TPropCell},
     },
-    prelude::Prop,
+    storage::timeindex::TimeIndexEntry,
 };
-use raphtory_api::core::entities::properties::meta::Meta;
 use rustc_hash::FxHashMap;
+
+use crate::LocalPOS;
 
 use super::properties::{PropEntry, Properties};
